Add sync method to RemoteUtils
This tries to pull followed by a push on a remote repository. When the
remote branch reference doesn't exists, the error is normalized using
`ERRORS.REF_NOT_FOUND`, which results in the error being ignored to
follow by a push.

Tests have been added as well to test the remaining of the RemoteUtils
methods (push, pull, list, edit).